bluetooth: Synchronize AVRCP Absolute Volume with A2DP sink
Like the previous commit this handles `Volume` property changes but
applies them to an a2dp sink instead of source stream. Only real_volume
is updated to make sure the stream doesn't apply any attenuation (using
software PCM) since this already happens on the sink side.
In addition to receiving volume events it is also possible to control
this remote volume slider by writing a new value to the DBus property.
This is especially useful when playing back to in-ear audio devices
which usually lack physical buttons to adjust the final volume on the
sink.