FIXUP: Ignore .Set result for Volume property
The only reason to asynchronously handle a response to this call is to
catch an error. A change notification for the Volume property is
received anyway, allowing us to determine what the actual applied volume
is.

In fact, that notification is received _before_ set_volume_reply is
handled: if bluez and/or the headphones report a change to a different
volume this will not be caught as the _reply handler overwrites it,
leading to more desyncs.